refactor: Add type_of_evaluation field to AddEvent form
The commit refactors the AddEvent form in the database/forms.py file to include the new type_of_evaluation field. This change aligns with the recent modification in the Event model, which added the type_of_evaluation field. By updating the form, we ensure that the user can input the type of evaluation when creating a new event.